Model_Deployment/requirement.txt
streamlit==1.44.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.4
matplotlib-inline==0.1.7
tensorflow==2.18.0
keras==3.10.0
pillow==11.2.1
imageio==2.37.0